// superset-frontend/src/cccs-viz/plugins/plugin-chart-at-a-glance/src/types.ts
export interface AtAGlanceGeoProps {
    ip: string;
    country: string;
    carrier: string;
    asn: string;
    organisation: string;
    connection_type: string;
    decimal: string;
    virusTotalCount: string; 
    city: string;
  }
  export interface AtAGlanceFarsightProps {
    associatedHostNames: string[];  
  }